+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Autoplay configuration
 * Readd server-side manipulation delete, upload and rename
+* Possibility to remove Icon, Size and Last Modified Date
 
 
 # Changelog